gateway/eelib/requirements.txt
certifi==2019.3.9
chardet==3.0.4
html-testRunner==1.2
idna==2.8
Jinja2==2.9.5
MarkupSafe==1.1.1
pyserial==3.4
requests==2.22.0
urllib3==1.25.3
